Make e fila

+++ Tp2/src/filaClau.h
@@ -9,11 +9,11 @@
 typedef int TChave;
 typedef struct TItem{
     TChave chave;
-    int linha;
+    long long int linha;
     int coluna;
     int valor;
 }TItem; 
-// é o item da lista, por exemplo: o item m[i][j] da matriz. Ou seja, deve conter o conteúdo.
+// Item da lista, por exemplo: o item m[i][j] da matriz. Ou seja, deve conter o conteúdo.
 // No exemplo acima (mat[i][j]) o conteúdo seria : linha, coluna e o valor da posição.
 
 typedef struct Celula* Apontador;
@@ -21,7 +21,7 @@ typedef struct Celula{
     TItem item;
     struct Celula* pProx;
 }TCelula;
-// tem como principal função "encapsular" o item, e gerar o apontador pro proximo item.
+// Tem como principal função "encapsular" o item, e gerar o apontador pro proximo item.
 
 typedef struct TFila{
     Apontador pFrente;
